% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\name{addInteractions}
\alias{addInteractions}
\title{Add interaction terms for each pair of putative variables}
\usage{
addInteractions(Z)
}
\arguments{
\item{Z}{a numeric matrix of putative variables, n (samples) by p (covariates).}
}
\value{
a matrix which includes the input matrix, Z, and all the two-way interactions.
}
\description{
Create all the second-order polynomials Z_i*Z_j where Z_i are columns of the matrix
Z (including Z_i*Z_i), and Z is an n (samples) by p (covariates)
}
\examples{
Z <- addInteractions(matrix(c(1,2,3, 11,12,13), nrow = 2,
ncol = 3, byrow = TRUE,dimnames = list(c("row1", "row2"),
  c("C.1", "C.2", "C.3"))))
}
\keyword{interactions}
