% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\name{fitSEMMS}
\alias{fitSEMMS}
\alias{SEMMS}
\title{Run the Generalized Alternating Maximization algorithm to fit the SEMMS model}
\usage{
fitSEMMS(
  dat,
  mincor = 0.7,
  nn = 5,
  nnset = c(),
  distribution,
  rnd = F,
  BHthr = 0.01,
  minchange = 1,
  maxst = 20,
  ptf = F,
  verbose = FALSE
)
}
\arguments{
\item{dat}{The dataset, as generated by readInputFile().}

\item{mincor}{The minimum correlation coefficient between pairs of putative variable, over which they are considered highly correlated. Default is 0.75.}

\item{nn}{The initial value for the number of non-null variables. Default is 5.}

\item{nnset}{Optional: instead of an initial number of candidates, can suggest specific column number in the Z matrix. Default is null.}

\item{distribution}{The distribution of the response (N, P, or B).}

\item{rnd}{Whether to run the greedy (F, default) algorithm, or the randomized.}

\item{BHthr}{The Benjamini-Hochberg threshold to be used when determining the initial set of non-null variables. Default=0.01.}

\item{minchange}{The minimum change in the log-likelihood that is to be considered meaningful. Default=1.}

\item{maxst}{The maximum number of iterations of the algorithm. Default=20.}

\item{ptf}{Whether to print output at each iteration to a file called SEMMS.log. Default=FALSE.}

\item{verbose}{Whether to show progress message to the user. Default=FALSE.}
}
\value{
\item{inits}{The initial values obtained from initVals().}
\item{initNN}{The column numbers of the putative variables to be included in the model in the first iteration.}
\item{gam.out}{The output from the GAMupdate() function, which is a list containing
nn=the selected variables, mu,beta,s2r,s2e=the parameter estimates of the mixture model,
gam_nn=the sign (+/-1) of the effect of the non-null variables, pp=a table of posterior probabilities,
and lockedOut=any variables found to be highly correlated with selected variables.}
\item{distribution}{The distribution selected by the user.}
\item{mincor}{The user's input for micor above.}
}
\description{
Coefficients of predictors in a GLM are modeled as a three-component normal
mixture, where the majority of predictors are assumed to belong to the null
component (i.e., they have no effect on the mean of the response) and the others can
have a positive, or negative effect. fitSEMMS runs a Generalized Alternating
Maximization algorithm to fit the model.
}
\examples{
\dontrun{
fn <- system.file("extdata", "AR1SIM.RData", package = "SEMMS", mustWork = TRUE)
dataYXZ <- readInputFile(fn, ycol=1, Zcols=2:100)
fittedSEMMS <- fitSEMMS(dataYXZ, mincor=0.8, nn=15, minchange= 1,
                        distribution="N",verbose=T,rnd=F)}
}
\keyword{Alternating}
\keyword{Generalized}
\keyword{Maximization}
\keyword{SEMMS}
\keyword{algorithm}
