% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GAMupdate}
\alias{GAMupdate}
\title{Run the GAM algorithm to select non-null variables}
\usage{
GAMupdate(
  initidx,
  initval,
  Yr,
  Xr,
  Z,
  distr = "N",
  randomize = TRUE,
  mincor = 0.7,
  maxsteps = 20L,
  minchange = 1,
  ptf = TRUE
)
}
\arguments{
\item{initidx}{an initial set of variables to use in the fitting algorithm}

\item{initval}{the values (-1 or 1) of the initial set of predictors}

\item{Yr}{the response vector}

\item{Xr}{the fixed-effect design matrix}

\item{Z}{the matrix of all putative variables}

\item{distr}{the GLM distribution to fit (N=Normal, B=binary, P=Poisson)}

\item{randomize}{Boolean - whether to run the greedy or randomized version}

\item{mincor}{a threshold of (absolute) correlation above which a pair is considered highly correlated}

\item{maxsteps}{maximum number of GAM iterations}

\item{minchange}{the minimum difference in log-likelihood between consecutive iterations below which we assume that the algorithm has converged}

\item{ptf}{Boolean - whether to print debug messages to SEMMS.log}
}
\value{
a list with elements: \code{nn} (1-based integer indices of selected
  non-null variables), \code{mu} (mixture mean), \code{beta} (mixture
  regression coefficients), \code{s2r} (random-effect variance), \code{s2e}
  (residual variance), \code{gam_nn} (sign, +/-1, of the effect of each
  selected variable), \code{pp} (matrix of posterior probabilities), and
  \code{lockedOut} (integer vector indicating variables excluded because they
  are highly correlated with a selected variable; non-zero value = index of
  the selected variable they are correlated with).
}
\description{
Run the GAM algorithm to select non-null variables
}
