% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initVals}
\alias{initVals}
\title{Get an initial set of putative variables for the GAM algorithm}
\usage{
initVals(Z, Yr, mincor = 0.7)
}
\arguments{
\item{Z}{matrix of putative variables}

\item{Yr}{the (possibly transformed) response}

\item{mincor}{a threshold of (absolute) correlation above which a pair is considered highly correlated}
}
\value{
a list with elements: \code{discard} (integer vector; for each
  variable, the index of the variable it is correlated with and therefore
  discarded in favour of, or -1 if kept) and \code{beta} (numeric vector of
  simple linear regression coefficients, one per column of Z).
}
\description{
Get an initial set of putative variables for the GAM algorithm
}
