% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\name{mds2D}
\alias{mds2D}
\title{Perform 2D multidimensional scaling}
\usage{
mds2D(dist1)
}
\arguments{
\item{dist1}{a distance matrix}
}
\description{
As in Diaconis et al. (2008). Returns a list named conf, with
two columns (x,y coordinates)
}
\examples{
\dontrun{plot(mds2D(as.matrix(dist(iris[,1:4])))$conf, pch=19, cex=0.7, col=4, axes=FALSE)}
}
\references{
Diaconis, Persi; Goel, Sharad; Holmes, Susan. Horseshoes in multidimensional scaling and local kernel methods. Ann. Appl. Stat. 2 (2008), no. 3, 777--807.
}
\keyword{multidimensional}
\keyword{scaling}
