% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\docType{data}
\name{NKI70}
\alias{NKI70}
\title{NKI70 data}
\format{
The NKI70 data has 144 observations and 72 predictors.
}
\description{
The NKI70 data is available from the “penalized” package in R (Goeman, 2010)
}
\examples{
 \dontrun{
fn <- system.file("extdata", "NKI70_t1.RData", package = "SEMMS", mustWork = TRUE)
dataYXZ <- readInputFile(fn, ycol=1, Zcols=2:73,addIntercept = TRUE)
fittedSEMMS <- fitSEMMS(dataYXZ, mincor=0.8, nn=6, minchange= 1,
  distribution="P", verbose=T, rnd=F)
fittedGLM <- runLinearModel(dataYXZ,fittedSEMMS$gam.out$nn, "P")
plotMDS(dataYXZ, fittedSEMMS, fittedGLM, ttl="NKI70")}
}
\references{
Goeman, J J. 2010. "L1 penalized estimation in the Cox proportional hazards model." no. 1. 14
}
\keyword{datasets}
