% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\docType{data}
\name{ozone}
\alias{ozone}
\title{Ozone data}
\format{
A text file with the ozone data. The data starts on line 20.
}
\description{
The air-pollution data set analyzed in this case study was first introduced in (Breiman and Friedman 1985). It consists of daily measurements of ozone concentration levels in the Los Angeles basin, collected over 330 days in 1976. There are eight meteorological explanatory variables, plut the day of the year. Note that the response (column 2) has to be log transformed. The predictos appear in columns 3-11. We add an intercept, and construct the two-way interaction terms in the example below.
}
\examples{
 \dontrun{
fn <- system.file("extdata", "ozone.txt", package = "SEMMS", mustWork = TRUE)
dataYXZ <- readInputFile(fn, ycol=2, skip=19, Zcols=3:11,
  addIntercept = TRUE, logTransform = 2, twoWay = TRUE)
fittedSEMMS <- fitSEMMS(dataYXZ, mincor=0.8, nn=15, minchange= 1,
                        distribution="N",verbose=T,rnd=F)
fittedGLM <- runLinearModel(dataYXZ,fittedSEMMS$gam.out$nn, "N")
plotMDS(dataYXZ, fittedSEMMS, fittedGLM, ttl="Ozone")}
}
\references{
Breiman, L, and J Friedman. 1985.
"Estimating Optimal Transformations for Multiple Regression and Correlation."
Technometrics 80: 580-598
}
\keyword{datasets}
