% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\name{plotFit}
\alias{plotFit}
\title{Diagnostics plots - fitted vs observed, and fitted vs. residuals}
\usage{
plotFit(fittedGLM, ttl = "")
}
\arguments{
\item{fittedGLM}{a list with 3 elements: mod = the glm or lm object, aic, and vif,
as generated by runLinearModel()}

\item{ttl}{An optional title for the plots (default is blank)}
}
\description{
Create two diagnostics plots for the fitted model
}
\examples{
\dontrun{
fn <- system.file("extdata", "AR1SIM.RData", package = "SEMMS", mustWork = TRUE)
dataYXZ <- readInputFile(fn, ycol=1, Zcols=2:100)
fittedSEMMS <- fitSEMMS(dataYXZ, mincor=0.8, nn=15, minchange= 1,
                        distribution="N",verbose=T,rnd=F)
fittedGLM <- runLinearModel(dataYXZ,fittedSEMMS$gam.out$nn, "N")
plotFit(fittedGLM)}
}
\keyword{Diagnostics}
\keyword{plots}
