% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot the fitted model found by SEMMS as a network.}
\usage{
plotMDS(dataYXZ, fittedSEMMS, fittedGLM, ttl = "")
}
\arguments{
\item{dataYXZ}{the SEMMS dataset, created by readInputFile()}

\item{fittedSEMMS}{the fitted model, returned by fitSEMMS()}

\item{fittedGLM}{a list with 3 elements: mod = the glm or lm object, aic, and vif,
as generated by runLinearModel()}

\item{ttl}{An optional title for the plot (default is blank)}
}
\description{
Plot a network of the response, Y, and all the variables that were found by SEMMS to be
associated with Y. If additional variables are highly correlated with ones found by
SEMMS, they will also be plotted, with edges between pairs of highly correlated
variables.
}
\examples{
\dontrun{
fn <- system.file("extdata", "AR1SIM.RData", package = "SEMMS", mustWork = TRUE)
dataYXZ <- readInputFile(fn, ycol=1, Zcols=2:100)
fittedSEMMS <- fitSEMMS(dataYXZ, mincor=0.8, nn=15, minchange= 1,
                        distribution="N",verbose=T,rnd=F)
fittedGLM <- runLinearModel(dataYXZ,fittedSEMMS$gam.out$nn, "N")
plotMDS(dataYXZ, fittedSEMMS, fittedGLM, ttl="AR1 simulation")}
}
\keyword{Network}
\keyword{diagram}
