% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\name{printModelParams}
\alias{printModelParams}
\title{Print the SEMMS model parameter estimates and non-null variables}
\usage{
printModelParams(gam.out)
}
\arguments{
\item{gam.out}{the element called gam.out in the list returned from fitSEMMS()}
}
\description{
Print the SEMMS model parameter estimates and non-null variables
}
\examples{
\dontrun{
#' fn <- system.file("extdata", "AR1SIM.RData", package = "SEMMS", mustWork = TRUE)
dataYXZ <- readInputFile(fn, ycol=1, Zcols=2:100)
fittedSEMMS <- fitSEMMS(dataYXZ, mincor=0.8, nn=15, minchange= 1,
                        distribution="N",verbose=T,rnd=F)
printModelParams(fittedSEMMS$gam.out) }
}
\keyword{estimates}
\keyword{parameter}
