% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\name{readInputFile}
\alias{readInputFile}
\title{Read the input data and creating the data structures}
\usage{
readInputFile(
  file,
  skip = 0,
  header = TRUE,
  sep = "\\t",
  ycol,
  Zcols,
  Xcols = c(),
  addIntercept = TRUE,
  logTransform = c(),
  twoWay = FALSE
)
}
\arguments{
\item{file}{The input file name.}

\item{skip}{How many lines to skip in the input file (default=0).}

\item{header}{Does the first line being read contain column names? (default=TRUE).}

\item{sep}{The delimiter (default is \\t).}

\item{ycol}{The number of the column in the input file which should be used as the response.}

\item{Zcols}{The columns in the input file which contain the putative variables.}

\item{Xcols}{The columns in the input file which contain the fixed effects in the model (default is none, c()).}

\item{addIntercept}{Whether to add an intercept column to the fixed effects (default=TRUE).}

\item{logTransform}{The column numbers which have to be log-transformed (default is none, c()).}

\item{twoWay}{Whether to add all two-way interactions to the putative variables (default=FALSE).}
}
\value{
\item{Z}{A numeric matrix of putative variables (the columns of Z)}
\item{Y}{The response vector}
\item{X}{Fixed effects matrix (could be null)}
\item{originalZnames}{The original names of the columns of the matrix Z}
\item{colnamesZ}{The generated column names, Z0001, Z0002, etc.}
\item{N}{The number of samples, nrow(Z)}
\item{K}{The number of putative variables, ncol(Z)}
\item{P}{The number of fixed effects, ncol(X)}
}
\description{
Read a text file containing the response variable, any fixed effects (possibly none),
and the putative variables. An intercept does not have to be in the data file - it
can be added by this function. The function can also add all two-way interactions.
The data may be loaded from an RData file, if is has been stored in the format that this
function generates.
The interaction terms are transformed using the scale function.
}
\examples{
\dontrun{
fn <- system.file("extdata", "ozone.txt", package = "SEMMS", mustWork = TRUE)
dataYXZ <- readInputFile(fn, ycol=2, Zcols=3:11,
addIntercept = TRUE, logTransform = 2, twoWay = TRUE)}
}
\keyword{data}
\keyword{format}
\keyword{input}
