% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\name{runLinearModel}
\alias{runLinearModel}
\title{Run the linear model using the selected variables}
\usage{
runLinearModel(dat, nnt, distr, func = "glm")
}
\arguments{
\item{dat}{The dataset, as generated by readInputFile().}

\item{nnt}{The column numbers in the Z matrix of the non-null variables.}

\item{distr}{The distribution to use when fitting the linear model (N=gaussian, P=poisson, or B=binomial).}

\item{func}{The function to use when fitting the model: glm or lm (default=glm).}
}
\value{
\item{Z}{A numeric matrix of putative variables (in the columns)}
\item{mod}{The object returned from glm or lm.}
\item{aic}{The AIC of the model.}
\item{vif}{The variance inflation factors (if more than one variable is used in the model).}
}
\description{
Run glm with the selected variables (identified by nnt) as predictors.
}
\examples{
\dontrun{
fn <- system.file("extdata", "AR1SIM.RData", package = "SEMMS", mustWork = TRUE)
dataYXZ <- readInputFile(fn, ycol=1, Zcols=2:100)
fittedSEMMS <- fitSEMMS(dataYXZ, mincor=0.8, nn=15, minchange= 1,
                        distribution="N",verbose=T,rnd=F)
fittedGLM <- runLinearModel(dataYXZ,fittedSEMMS$gam.out$nn, "N")}
}
\keyword{GLM}
