% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMMS-package.R
\docType{package}
\name{SEMMS-package}
\alias{SEMMS-package}
\title{SEMMS: Variable Selection in Generalized Linear Models via Empirical Bayes}
\description{
Fits a three-component normal mixture model to the coefficients in the linear predictor of a GLM (Gaussian, Poisson, or binomial response) using an empirical Bayes approach and a Generalized Alternating Maximization algorithm. Version 0.4.0 adds fitSEMMSmixed(), an alternating coordinate-ascent extension that incorporates subject-level random intercepts and/or random slopes via lme4, enabling variable selection in mixed-effects models with high-dimensional predictor sets.
}
\author{
\strong{Maintainer}: Haim Bar \email{haim.bar@uconn.edu}

}
\keyword{internal}
