% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\docType{data}
\name{SimBin}
\alias{SimBin}
\title{Simulated gene expression data with a binary response}
\format{
A 1001 by 120 matrix.
}
\description{
SimBin is a simulated dataset with an autoregressive structure. The true predictors
are Z1-Z5 and Z101-Z105. Z1-Z5 have an AR(1) structure with rho=0.95, and so do Z101-Z105.
}
\examples{
\dontrun{
fn <- system.file("extdata", "SimBin.RData", package = "SEMMS", mustWork = TRUE)
dataYXZ <- readInputFile(fn, ycol=1, Zcols=2:1001)
fittedSEMMS <- fitSEMMS(dataYXZ, mincor=0.7, nn=5, minchange  = 1,
     distribution="B", verbose=T, rnd=F)
fittedGLM <- runLinearModel(dataYXZ,fittedSEMMS$gam.out$nn, "B")
plotMDS(dataYXZ, fittedSEMMS, fittedGLM, ttl="Simulated Binomial (AR1)")}
}
\keyword{datasets}
