% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initVals}
\alias{initVals}
\title{Get an initial set of putative variables for the GAM algorithm}
\usage{
initVals(Z, Yr, mincor = 0.7)
}
\arguments{
\item{Z}{matrix of putative variables}

\item{Yr}{the (possibly transformed) response}

\item{mincor}{a threshold of (absolute) correlation above which a pair is considered highly correlated}
}
\value{
a list containing  variables to ignore because they are highly correlated with other, and SLR coefficients
}
\description{
Get an initial set of putative variables for the GAM algorithm
}
