% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmsFuncs.R
\name{runMixedModel}
\alias{runMixedModel}
\title{Run the mixed model using variables selected by fitSEMMSmixed}
\usage{
runMixedModel(dat, nnt, re_term)
}
\arguments{
\item{dat}{Dataset created by readInputFile() (with group_col set).}

\item{nnt}{Integer vector of selected column indices in Z (from gam.out$nn).}

\item{re_term}{Character string giving the lme4 random-effects term, as
returned in fitSEMMSmixed()$re_term.}
}
\value{
A list with elements:
\item{mod}{The lmer model object.}
\item{aic}{AIC of the fitted model.}
\item{vif}{Variance inflation factors for the fixed effects (NA if only one predictor).}
}
\description{
Convenience wrapper that refits the final REML lmer with the variable
indices in nnt, mirroring the interface of runLinearModel().
}
\keyword{lmer}
\keyword{mixed}
\keyword{model}
